--- src/app/api/auth/[...nextauth]/route.ts
@@ -1,25 +1,0 @@
-import NextAuth from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-
-const mySecret = "balolebwamitchasingathebest";
-
-const handler = NextAuth({
-  // Configure one or more authentication providers
-  secret: process.env.NEXTAUTH_SECRET || mySecret,
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      id: "credentials",
-      credentials: {
-        username: { label: "enter your email here", type: "email", placeholder: "tchasingajacque@gmail.com" },
-        password: { label: "enter your password here", type: "password" }
-      },
-      async authorize(credentials, req) {
-        console.log({ credentials });
-        return null;
-      }
-    })
-  ]
-});
-
-export { handler as GET, handler as POST };
